fix(foc-pr-report): move client into package, fix f-string syntax, fix unknown-repo links
- Move foc_project14_client.py into foc_pr_report/ so the built package
  is self-contained and installable without the repo checkout
- Remove sys.path mutation from cli.py; use absolute package imports instead
- Update foc_wg_pr_notifier.py to import from foc_pr_report.foc_project14_client
  via a sys.path insert pointing at foc-pr-report/
- Fix SyntaxError on Python <3.12: backslash inside nested f-string expression
  in render_repo_status_markdown (report.py line 194)
- Fix invalid repo:unknown filter in non-zero unknown-repo cells (report.py line 223)
